Moved integration tests to separate package
